Put interface Logtailer and its implementation together;
